// types/juz.types.ts
export type JuzData = {
	ayat_arab: string;
	ayat_indo: string;
	ayat_latin: string;
	name: string;
	name_end_arab: string;
	name_end_id: string;
	name_start_arab: string;
	name_start_id: string;
	number: string;
	surah_id_end: string;
	surah_id_start: string;
	verse_end: string;
	verse_start: string;
};